minimap implemented

+++ libft/string/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ychibani <ychibani@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/22 17:15:25 by ychibani          #+#    #+#             */
-/*   Updated: 2022/09/10 12:48:06 by ychibani         ###   ########.fr       */
+/*   Updated: 2022/01/16 15:35:55 by ychibani         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	__strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (!s)
